+++ examples/demo.rs
@@ -5,7 +5,7 @@ struct Stage {
     show_egui_demo_windows: bool,
     egui_demo_windows: egui_demo_lib::DemoWindows,
     color_test: egui_demo_lib::ColorTest,
-    scale: f32,
+    pixels_per_point: f32,
 }
 
 impl Stage {
@@ -15,7 +15,7 @@ impl Stage {
             show_egui_demo_windows: true,
             egui_demo_windows: Default::default(),
             color_test: Default::default(),
-            scale: 1.0,
+            pixels_per_point: ctx.dpi_scale(),
         }
     }
 }
@@ -23,14 +23,15 @@ impl Stage {
 impl mq::EventHandler for Stage {
     fn update(&mut self, _ctx: &mut mq::Context) {}
 
-    fn draw(&mut self, ctx: &mut mq::Context) {
-        self.egui_mq.egui_input().pixels_per_point = Some(self.scale.clamp(0.7, 2.0));
-        ctx.clear(Some((1., 1., 1., 1.)), None, None);
-        ctx.begin_default_pass(mq::PassAction::clear_color(0.0, 0.0, 0.0, 1.0));
-        ctx.end_render_pass();
+    fn draw(&mut self, mq_ctx: &mut mq::Context) {
+        mq_ctx.clear(Some((1., 1., 1., 1.)), None, None);
+        mq_ctx.begin_default_pass(mq::PassAction::clear_color(0.0, 0.0, 0.0, 1.0));
+        mq_ctx.end_render_pass();
+
+        let dpi_scale = mq_ctx.dpi_scale();
 
         // Run the UI code:
-        self.egui_mq.run(ctx, |egui_ctx| {
+        self.egui_mq.run(mq_ctx, |_mq_ctx, egui_ctx| {
             if self.show_egui_demo_windows {
                 self.egui_demo_windows.ui(egui_ctx);
             }
@@ -39,13 +40,18 @@ impl mq::EventHandler for Stage {
                 egui::widgets::global_dark_light_mode_buttons(ui);
                 ui.checkbox(&mut self.show_egui_demo_windows, "Show egui demo windows");
 
-                if ui.button("Increase scale").clicked() {
-                    self.scale += 0.1;
-                }
-
-                if ui.button("Decrease scale").clicked() {
-                    self.scale -= 0.1;
-                }
+                ui.group(|ui| {
+                    ui.label("Physical pixels per each logical 'point':");
+                    ui.label(format!("native: {:.2}", dpi_scale));
+                    ui.label(format!("egui:   {:.2}", ui.ctx().pixels_per_point()));
+                    ui.add(
+                        egui::Slider::new(&mut self.pixels_per_point, 0.75..=3.0).logarithmic(true),
+                    )
+                    .on_hover_text("Physical pixels per logical point");
+                    if ui.button("Reset").clicked() {
+                        self.pixels_per_point = dpi_scale;
+                    }
+                });
 
                 #[cfg(not(target_arch = "wasm32"))]
                 {
@@ -55,7 +61,16 @@ impl mq::EventHandler for Stage {
                 }
             });
 
+            // Don't change scale while dragging the slider
+            if !egui_ctx.is_using_pointer() {
+                egui_ctx.set_pixels_per_point(self.pixels_per_point);
+            }
+
             egui::Window::new("Color Test").show(egui_ctx, |ui| {
+                ui.colored_label(
+                    egui::Color32::RED,
+                    "NOTE: egui-miniquad does not currently pass the color test!",
+                );
                 egui::ScrollArea::both()
                     .auto_shrink([false; 2])
                     .show(ui, |ui| {
@@ -66,11 +81,11 @@ impl mq::EventHandler for Stage {
 
         // Draw things behind egui here
 
-        self.egui_mq.draw(ctx);
+        self.egui_mq.draw(mq_ctx);
 
         // Draw things in front of egui here
 
-        ctx.commit_frame();
+        mq_ctx.commit_frame();
     }
 
     fn mouse_motion_event(&mut self, _: &mut mq::Context, x: f32, y: f32) {
@@ -129,6 +144,8 @@ impl mq::EventHandler for Stage {
 fn main() {
     let conf = mq::conf::Conf {
         high_dpi: true,
+        window_width: 1200,
+        window_height: 1024,
         ..Default::default()
     };
     mq::start(conf, |mut ctx| Box::new(Stage::new(&mut ctx)));