+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
+* You can now scale the GUI with `egui_ctx.set_pixels_per_point(…)`.
+* Add miniquad context as a parameter in the lambda in `EguiMq::run`
 
 # 0.11.0 - 2022-05-01
 * Update egui to `0.18.1`